create and merge in the same step
