+++ src/OSPSuite.Core/Domain/Services/EventGroupMerger.cs
@@ -1,12 +1,10 @@
 ﻿using System;
-using System.Collections;
 using System.Linq;
 using OSPSuite.Assets;
 using OSPSuite.Core.Domain.Builder;
 using OSPSuite.Core.Domain.Mappers;
 using OSPSuite.Utility.Exceptions;
 using OSPSuite.Utility.Extensions;
-using System.Reflection;
 using OSPSuite.Core.Domain.Descriptors;
 using OSPSuite.Core.Extensions;
 using OSPSuite.Utility.Collections;
@@ -16,135 +14,178 @@ namespace OSPSuite.Core.Domain.Services
 {
    internal interface IEventGroupMerger
    {
-      IContainer MergeEventGroupsContainer(ModelConfiguration modelConfiguration);
+      void MergeEventGroupsContainer(ModelConfiguration modelConfiguration);
    }
 
    internal class EventGroupMerger : IEventGroupMerger
    {
       private readonly IEventGroupBuilderToEventGroupMapper _eventGroupMapper;
       private readonly IContainerMergeTask _containerMergeTask;
-      private readonly IEventBuilderTask _eventBuilderTask;
+      private readonly IKeywordReplacerTask _keywordReplacerTask;
+      private readonly ITransportBuilderToTransportMapper _transportMapper;
 
       public EventGroupMerger(
          IContainerMergeTask containerMergeTask,
-         IEventBuilderTask eventBuilderTask,
-         IEventGroupBuilderToEventGroupMapper eventGroupMapper)
+         IEventGroupBuilderToEventGroupMapper eventGroupMapper, IKeywordReplacerTask keywordReplacerTask, ITransportBuilderToTransportMapper transportMapper)
       {
          _eventGroupMapper = eventGroupMapper;
+         _keywordReplacerTask = keywordReplacerTask;
+         _transportMapper = transportMapper;
          _containerMergeTask = containerMergeTask;
-         _eventBuilderTask = eventBuilderTask;
       }
 
-      public IContainer MergeEventGroupsContainer(ModelConfiguration modelConfiguration)
+      public void MergeEventGroupsContainer(ModelConfiguration modelConfiguration)
       {
-         var (model, simulationBuilder) = modelConfiguration;
-
-         _eventBuilderTask.CreateEvents(modelConfiguration);
-         //When runniong the should_not_crash test:
-            //we have the "Bolus Application" here -> Root|Organism|ArterialBlood|Plasma|Bolus Application
-         createMergedContainerStructureInRoot(model.Root, simulationBuilder);
-         //After this it is on the Root Container, due to this line inside the createMergedContainerStructureInRoot method:
-         //root.AddChildren(firstEventGroupBuildingBlock.Select(mapToModelContainer));
-
-         return model.Root;
+         createMergedContainerStructureInRoot(modelConfiguration);
       }
       private ICache<DescriptorCriteria, IEnumerable<IContainer>> _sourceCriteriaTargetContainerCache;
-      private void createMergedContainerStructureInRoot(IContainer root, SimulationBuilder simulationBuilder)
+      private Cache<DescriptorCriteria, IEnumerable<IContainer>> _applicationTransportTargetContainerCache;
+      private EntityDescriptorMapList<IContainer> _allModelContainerDescriptors;
+
+      private void createMergedContainerStructureInRoot(ModelConfiguration modelConfiguration)
       {
-         _sourceCriteriaTargetContainerCache = new Cache<DescriptorCriteria, IEnumerable<IContainer>>();
-         var allEventGroupAndMergeBehaviors = simulationBuilder.EventGroupAndMergeBehaviors;
-         if (!allEventGroupAndMergeBehaviors.Any())
-            return;
+         try
+         {
+            var (model, simulationBuilder) = modelConfiguration;
 
-         var firstEventGroupBuildingBlock = allEventGroupAndMergeBehaviors[0].eventGroupBuildingBlock;
-         var allOtherSpatialStructuresWithMergeBehavior = allEventGroupAndMergeBehaviors.Skip(1).ToList();
-         var mapToModelContainer = mapContainerDef(simulationBuilder);
+            _allModelContainerDescriptors = model.Root.GetAllContainersAndSelf<IContainer>().ToEntityDescriptorMapList();
+            _sourceCriteriaTargetContainerCache = new Cache<DescriptorCriteria, IEnumerable<IContainer>>();
+            _applicationTransportTargetContainerCache = new Cache<DescriptorCriteria, IEnumerable<IContainer>>();
+            var allEventGroupAndMergeBehaviors = simulationBuilder.EventGroupAndMergeBehaviors;
 
+            allEventGroupAndMergeBehaviors.Each(eventGroupBuildingBlockAndMerge =>
+            {
+               var (buildingBlock, mergeBehavior) = eventGroupBuildingBlockAndMerge;
+               mergeEventGroups(modelConfiguration, buildingBlock, mergeBehavior);
+            });
+         }
+         finally
+         {
+            _allModelContainerDescriptors = null;
+            _sourceCriteriaTargetContainerCache.Clear();
+            _sourceCriteriaTargetContainerCache = null;
+            _applicationTransportTargetContainerCache.Clear();
+            _applicationTransportTargetContainerCache = null;
+         }
+      }
 
-         //Adding this to replicate somehow, how the containers are added to the model
-         //This is extracted from the CreateEvents method in the EventBuilderTask class
-         //Cache all containers where the event group builder will be created using the source criteria
-         //Depending on the test running, the should_have_merged_the_events_with_second_present_only or should_not_crash
-         //we have a different outcome from this. Again this is just a test that I`m doing.
-
-         //I think we might have to find what container to place the events on by this criteria somehow.
-         //as mentioned befor, it is setting them on "PLASMA" container. But what if we don`t have one?
-         //that`s happening with my should_have_merged_the_events_with_second_present_only  test, maybe I 
-         //should add it to the construction of the modules? 
-
-         EntityDescriptorMapList<IContainer> _allModelContainerDescriptors;
-         _allModelContainerDescriptors = root.GetAllContainersAndSelf<IContainer>().ToEntityDescriptorMapList();
-
-         foreach (var eventGroupBuilder in simulationBuilder.EventGroups)
+      private void mergeEventGroups(ModelConfiguration modelConfiguration, EventGroupBuildingBlock buildingBlock, MergeBehavior mergeBehavior)
+      {
+         foreach (var eventGroupBuilder in buildingBlock)
          {
             if (_sourceCriteriaTargetContainerCache.Contains(eventGroupBuilder.SourceCriteria))
                continue;
 
             _sourceCriteriaTargetContainerCache.Add(eventGroupBuilder.SourceCriteria, _allModelContainerDescriptors.AllSatisfiedBy(eventGroupBuilder.SourceCriteria));
          }
-         //This contains all the containers to add the eventgroups? -> _sourceCriteriaTargetContainerCache
 
-         foreach (var eventGroupBuilder in simulationBuilder.EventGroups)
-         {
-            var lst = _allModelContainerDescriptors.AllSatisfiedBy(eventGroupBuilder.SourceCriteria);
-         }
-         
-
-
-         root.AddChildren(firstEventGroupBuildingBlock.Select(mapToModelContainer));
-
-         //var children = firstEventGroupBuildingBlock.Select(mapToModelContainer);
-         //var uniqueChildren = children.Where(newChild => !root.Children.Any(existingChild => existingChild.Name == newChild.Name));
-         //root.AddChildren(uniqueChildren);
-
-         allOtherSpatialStructuresWithMergeBehavior
-            .Select(x => new { MergeBehavior = x.mergeBehavior, eventGroups = x.eventGroupBuildingBlock.Select(mapToModelContainer), buildingBlock = x.eventGroupBuildingBlock })
-            .Each(x => x.eventGroups.Each(eventGroup => tryMergeEventGroupInContainer(root, eventGroup, x.MergeBehavior, x.buildingBlock)));
+         buildingBlock.Each(x => createEventGroupFrom(x, modelConfiguration, buildingBlock, mergeBehavior));
       }
 
-      private void tryMergeEventGroupInContainer(IContainer root, IContainer eventGroup, MergeBehavior mergeBehavior, EventGroupBuildingBlock eventGroupBuildingBlock)
+      /// <summary>
+      ///    Adds event group to all model containers with defined criteria
+      /// </summary>
+      private void createEventGroupFrom(EventGroupBuilder eventGroupBuilder, ModelConfiguration modelConfiguration, EventGroupBuildingBlock eventGroupBuildingBlock, MergeBehavior mergeBehavior)
       {
+         foreach (var sourceContainer in _sourceCriteriaTargetContainerCache[eventGroupBuilder.SourceCriteria])
+         {
+            createEventGroupInContainer(eventGroupBuilder, sourceContainer, modelConfiguration, eventGroupBuildingBlock, mergeBehavior);
+         }
+      }
+
+      /// <summary>
+      ///    Adds event group to source container where event takes place
+      /// </summary>
+      private void createEventGroupInContainer(EventGroupBuilder eventGroupBuilder, IContainer sourceContainer, ModelConfiguration modelConfiguration, EventGroupBuildingBlock eventGroupBuildingBlock, MergeBehavior mergeBehavior)
+      {
+         //this creates recursively all event groups for the given builder
+         var (_, simulationBuilder, replacementContext) = modelConfiguration;
+         var eventGroup = _eventGroupMapper.MapFrom(eventGroupBuilder, simulationBuilder);
+         
+         tryMergeEventGroupInContainer(sourceContainer, eventGroup, mergeBehavior, eventGroupBuildingBlock);
+
+         //needs to add the required transport into model only for the added event group
+         foreach (var childEventGroup in eventGroup.GetAllContainersAndSelf<EventGroup>())
+         {
+            var childEventGroupBuilder = simulationBuilder.BuilderFor(childEventGroup).DowncastTo<EventGroupBuilder>();
+            if (childEventGroupBuilder is ApplicationBuilder applicationBuilder)
+               addApplicationTransports(applicationBuilder, childEventGroup, modelConfiguration);
+
+            _keywordReplacerTask.ReplaceIn(childEventGroup, childEventGroupBuilder, replacementContext);
+         }
+      }
+
+      private void addApplicationTransports(ApplicationBuilder applicationBuilder, EventGroup eventGroup, ModelConfiguration modelConfiguration)
+      {
+         var allEventGroupParentChildContainers = eventGroup.GetAllContainersAndSelf<IContainer>().ToEntityDescriptorMapList();
+         foreach (var appTransport in applicationBuilder.Transports)
+         {
+            var transportBuilder = appTransport;
+            if (!_applicationTransportTargetContainerCache.Contains(transportBuilder.TargetCriteria))
+               _applicationTransportTargetContainerCache.Add(appTransport.TargetCriteria, _allModelContainerDescriptors.AllSatisfiedBy(transportBuilder.TargetCriteria));
+
+            addApplicationTransportToModel(transportBuilder, allEventGroupParentChildContainers, applicationBuilder.MoleculeName, modelConfiguration);
+         }
+      }
+
+      private void addApplicationTransportToModel(TransportBuilder appTransport, EntityDescriptorMapList<IContainer> allEventGroupParentChildContainers, string moleculeName, ModelConfiguration modelConfiguration)
+      {
+         var appTransportSourceContainers = sourceContainersFor(appTransport, allEventGroupParentChildContainers);
+         var appTransportTargetContainers = _applicationTransportTargetContainerCache[appTransport.TargetCriteria].ToList();
+         var (_, simulationBuilder, replacementContext) = modelConfiguration;
+
+         foreach (var sourceContainer in appTransportSourceContainers)
+         {
+            var sourceAmount = sourceContainer.GetSingleChildByName<MoleculeAmount>(moleculeName);
+            if (sourceAmount == null)
+               throw new OSPSuiteException(Validation.CannotCreateApplicationSourceNotFound(appTransport.Name, moleculeName, sourceContainer.Name));
+
+            foreach (var targetContainer in appTransportTargetContainers)
+            {
+               var targetAmount = targetContainer.GetSingleChildByName<MoleculeAmount>(moleculeName);
+               if (targetAmount == null)
+                  throw new OSPSuiteException(Validation.CannotCreateApplicationTargetNotFound(appTransport.Name, moleculeName, targetContainer.Name));
+
+               var transport = _transportMapper.MapFrom(appTransport, simulationBuilder);
+
+               transport.SourceAmount = sourceAmount;
+               transport.TargetAmount = targetAmount;
+
+               _keywordReplacerTask.ReplaceIn(transport, moleculeName, replacementContext);
+
+               //At the moment, no neighborhoods between application sub-containers and
+               //spatial structure sub-containers are defined. Application transports are
+               //added as direct children of the source molecule amount
+               if (!sourceAmount.ContainsName(transport.Name))
+                  sourceAmount.Add(transport);
+               else
+                  throw new OSPSuiteException(Validation.TransportAlreadyCreatorForMolecule(appTransport.Name, transport.Name, moleculeName));
+            }
+         }
+      }
+
+      private IEnumerable<IContainer> sourceContainersFor(TransportBuilder transport, EntityDescriptorMapList<IContainer> allEventGroupParentChildContainers)
+      {
+         return allEventGroupParentChildContainers.AllSatisfiedBy(transport.SourceCriteria);
+      }
+
+      private void tryMergeEventGroupInContainer(IContainer mergeContainer, EventGroup eventGroup, MergeBehavior mergeBehavior, EventGroupBuildingBlock eventGroupBuildingBlock)
+      {
+         //probably should never happen
+         if (mergeContainer == null)
+            return;
+
          try
          {
-            mergeTopContainerInEventGroup(eventGroup, root, mergeBehavior);
+            if (mergeBehavior == MergeBehavior.Extend)
+               _containerMergeTask.AddOrMergeContainer(mergeContainer, eventGroup);
+            else
+               _containerMergeTask.AddOrReplaceInContainer(mergeContainer, eventGroup);
          }
          catch (Exception)
          {
-            throw new OSPSuiteException(Error.CannotFindParentContainerWithPath(root.ParentPath.PathAsString, root.Name, eventGroupBuildingBlock.Name, eventGroupBuildingBlock.Module.Name));
+            throw new OSPSuiteException(Error.CannotFindParentContainerWithPath(mergeContainer.ParentPath.PathAsString, mergeContainer.Name, eventGroupBuildingBlock.Name, eventGroupBuildingBlock.Module.Name));
          }
-      }
-
-      private Func<EventGroupBuilder, IContainer> mapContainerDef(SimulationBuilder simulationBuilder) => container => _eventGroupMapper.MapFrom(container, simulationBuilder);
-
-      private void mergeTopContainerInEventGroup(IContainer eventGroup, IContainer root, MergeBehavior mergeBehavior)
-      {
-         //probably should never happen
-         if (root == null)
-            return;
-
-         //In this case, we add or replace the top container
-         if (root.ParentPath == null || string.IsNullOrEmpty(root.ParentPath.PathAsString))
-            replaceOrMergeContainerIntoParent(root, eventGroup, mergeBehavior);
-         else
-            insertTopContainerIntoStructure(root, eventGroup, mergeBehavior);
-      }
-
-      private void insertTopContainerIntoStructure(IContainer topContainer, IContainer root, MergeBehavior mergeBehavior)
-      {
-         var parentContainer = root.ParentPath.Resolve<IContainer>(root);
-         if (parentContainer == null)
-            throw new Exception();
-
-         replaceOrMergeContainerIntoParent(parentContainer, topContainer, mergeBehavior);
-      }
-
-      private void replaceOrMergeContainerIntoParent(IContainer parentContainer, IContainer containerToMerge, MergeBehavior mergeBehavior)
-      {
-         //Merge behavior is extend or we have a special case to deal with when dealing with MoleculeProperties container that we merge instead of replacing
-         if (mergeBehavior == MergeBehavior.Extend)
-            _containerMergeTask.AddOrMergeContainer(parentContainer, containerToMerge);
-         else
-            _containerMergeTask.AddOrReplaceInContainer(parentContainer, containerToMerge);
       }
    }
 }
